# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowDelta API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "summary": "List Jobs",
        "operationId": "list_jobs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit",
        "operationId": "create_job",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads": {
      "post": {
        "summary": "Upload",
        "operationId": "upload_video",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "summary": "Detail",
        "operationId": "get_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/retry": {
      "post": {
        "summary": "Retry",
        "operationId": "retry_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/note": {
      "get": {
        "summary": "Note",
        "operationId": "get_note",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/frames/{frame_id}": {
      "get": {
        "summary": "Frame",
        "operationId": "get_frame",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "frame_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Frame Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/video": {
      "get": {
        "summary": "Video",
        "operationId": "get_video",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/download/{kind}": {
      "get": {
        "summary": "Download",
        "operationId": "download_note",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "zip",
                "md",
                "html"
              ],
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_video": {
        "properties": {
          "video": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Video"
          },
          "subtitles": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitles"
          },
          "mode": {
            "type": "string",
            "enum": [
              "extractive",
              "llm",
              "vision"
            ],
            "title": "Mode",
            "default": "extractive"
          },
          "asr_prompt": {
            "type": "string",
            "maxLength": 500,
            "title": "Asr Prompt",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "video"
        ],
        "title": "Body_upload_video"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Illustration": {
        "properties": {
          "frame_id": {
            "type": "string",
            "title": "Frame Id"
          },
          "caption": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Caption"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "frame_id",
          "caption"
        ],
        "title": "Illustration"
      },
      "JobCreate": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 10,
            "title": "Url"
          },
          "mode": {
            "type": "string",
            "enum": [
              "extractive",
              "llm",
              "vision"
            ],
            "title": "Mode",
            "default": "extractive"
          },
          "language": {
            "type": "string",
            "pattern": "^[a-z]{2,3}$",
            "title": "Language",
            "default": "zh"
          },
          "asr_model": {
            "type": "string",
            "enum": [
              "base",
              "small",
              "medium"
            ],
            "title": "Asr Model",
            "default": "small"
          },
          "asr_prompt": {
            "type": "string",
            "maxLength": 500,
            "title": "Asr Prompt",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "JobCreate"
      },
      "JobList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "JobList"
      },
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "extractive",
              "llm",
              "vision"
            ],
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "section_count": {
            "type": "integer",
            "title": "Section Count"
          },
          "image_count": {
            "type": "integer",
            "title": "Image Count"
          },
          "cover_frame_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Frame Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_url",
          "source_name",
          "mode",
          "status",
          "stage",
          "progress",
          "attempt",
          "error",
          "author",
          "duration_ms",
          "section_count",
          "image_count",
          "cover_frame_id",
          "created_at",
          "updated_at"
        ],
        "title": "JobView"
      },
      "Note": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "enum": [
              "extractive",
              "llm",
              "vision"
            ],
            "title": "Mode"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "minItems": 1,
            "title": "Sections"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "mode",
          "sections"
        ],
        "title": "Note"
      },
      "NoteFrame": {
        "properties": {
          "frame_id": {
            "type": "string",
            "title": "Frame Id"
          },
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms"
          }
        },
        "type": "object",
        "required": [
          "frame_id",
          "timestamp_ms"
        ],
        "title": "NoteFrame"
      },
      "NoteView": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobView"
          },
          "note": {
            "$ref": "#/components/schemas/Note"
          },
          "transcript": {
            "$ref": "#/components/schemas/Transcript"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/NoteFrame"
            },
            "type": "array",
            "title": "Frames"
          },
          "video_available": {
            "type": "boolean",
            "title": "Video Available"
          }
        },
        "type": "object",
        "required": [
          "job",
          "note",
          "transcript",
          "frames",
          "video_available"
        ],
        "title": "NoteView"
      },
      "Paragraph": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Text"
          },
          "source_segment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Source Segment Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "source_segment_ids"
        ],
        "title": "Paragraph"
      },
      "Section": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "paragraphs": {
            "items": {
              "$ref": "#/components/schemas/Paragraph"
            },
            "type": "array",
            "maxItems": 150,
            "minItems": 1,
            "title": "Paragraphs"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/Illustration"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Images"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "paragraphs"
        ],
        "title": "Section"
      },
      "Segment": {
        "properties": {
          "segment_id": {
            "type": "string",
            "title": "Segment Id"
          },
          "start_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "End Ms"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "source": {
            "type": "string",
            "enum": [
              "subtitle",
              "asr"
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "segment_id",
          "start_ms",
          "end_ms",
          "text",
          "source"
        ],
        "title": "Segment"
      },
      "Transcript": {
        "properties": {
          "segments": {
            "items": {
              "$ref": "#/components/schemas/Segment"
            },
            "type": "array",
            "minItems": 1,
            "title": "Segments"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "segments",
          "language",
          "origin"
        ],
        "title": "Transcript"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}